requirements.txt
beautifulsoup4>=4.12
icalendar>=5.0
requests>=2.31
playwright>=1.40
caldav>=1.3
python-telegram-bot>=21.0
cryptography>=42.0
